refactor(backend): give the apply response shapes their own module
The three field descriptions the schema-docs gate asked for pushed bots/schemas.py
to 1007 lines, past the 1000-line cap. Splitting is the answer the gate asks for
and the right one here: the apply report is four nested models with a worked
example, a cohesive concern that reads better beside the routes returning it
than buried in the shared bot schema module.

Pure move — the four classes are unchanged, and the two modules that used them
now import from the new one.
